+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/HighPriceIndicator.java
@@ -49,6 +49,7 @@ public class HighPriceIndicator extends AbstractIndicator<Num> {
         return getBarSeries().getBar(index).getHighPrice();
     }
 
+    /** @return {@code 0} */
     @Override
     public int getUnstableBars() {
         return 0;